feat(cosec): rename generator to openapi and update related packages
- Rename generatorSupport to openapiSupport in build.gradle.kts
- Update package names from cosec.generator to cosec.openapi.generator
- Move OpenAPI related classes to new package structure
- Update dependencies and imports to reflect new package names
