+++ cosec-openapi/src/main/kotlin/me/ahoo/cosec/openapi/generator/OpenAPIAppPermissionGenerator.kt
@@ -11,7 +11,7 @@
  * limitations under the License.
  */
 
-package me.ahoo.cosec.generator
+package me.ahoo.cosec.openapi.generator
 
 import io.swagger.v3.oas.models.OpenAPI
 import me.ahoo.cosec.api.permission.AppPermission
@@ -22,6 +22,7 @@ import me.ahoo.cosec.permission.PermissionData
 import me.ahoo.cosec.permission.PermissionGroupData
 import me.ahoo.cosec.policy.action.ACTION_MATCHER_METHOD_KEY
 import me.ahoo.cosec.policy.action.PathActionMatcherFactory
+import kotlin.collections.iterator
 
 object OpenAPIAppPermissionGenerator {
     private const val APP_ID = "AppId"
