+++ cosec-openapi/src/test/kotlin/me/ahoo/cosec/generator/OpenAPIPolicyGeneratorTest.kt
@@ -4,6 +4,7 @@ import io.swagger.v3.oas.models.OpenAPI
 import io.swagger.v3.oas.models.Operation
 import io.swagger.v3.oas.models.PathItem
 import io.swagger.v3.oas.models.Paths
+import me.ahoo.cosec.openapi.generator.OpenAPIPolicyGenerator
 import org.hamcrest.MatcherAssert.*
 import org.hamcrest.Matchers.*
 import org.junit.jupiter.api.Test
