+++ settings.gradle.kts
@@ -28,4 +28,4 @@ include(":cosec-gateway-server")
 include(":cosec-opentelemetry")
 include(":cosec-ip2region")
 include(":code-coverage-report")
-include(":cosec-generator")
+include(":cosec-openapi")
